+++ packages/wallet-sdk/src/sign/SignHandler.test.ts
@@ -1,4 +1,4 @@
-import { SignHandler } from './SignerConfigurator';
+import { SignHandler } from './SignHandler';
 
 const mockGetItem = jest.fn();
 jest.mock(':core/storage/ScopedLocalStorage', () => {
